Completed all assignments in the flex section.
Ensured all layouts matched the desired-outcome. Kept close to the
specification in terms of the CSS that I was required to utilise. However,
a number of my solutions were not as concise as the ones provided.
Utilised a few hacks. Sufficient for now, but still much to be learnt.

 # Please enter the commit message for your changes. Lines
starting